refactor: merge gui into main batt binary
